fix problem with the pyinstaller build on linux

+++ pylibcklb/installer/pyinstaller.py
@@ -11,6 +11,8 @@
 #
 # @bug          No bugs at the moment.
 #
+# @note         Some parts of this code comes original from Hermann Krumrey: gitlab-build-scripts-0.6.0, but was deleted from his server
+#
 # @warning      Warning this code is not tested until now!!!
 #
 # @copyright    pylibcklb package
@@ -30,63 +32,36 @@
 #               along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 import os
+import platform
 from subprocess import Popen
 from pylibcklb.metadata import PackageVariables
 from pylibcklb.ClassLibrary import cDebug
 Debug = cDebug(PackageVariables.DebugLevel)
 
-class PyInstaller:
-    
-    #note Code comes original from Hermann Krumrey: gitlab-build-scripts-0.6.0, but was deleted from his server
-    def BuildOnWindows(build_path, project_name, version_number):
-        main_py     = os.path.join(project_name, "main.py")
-        icon_file   = os.path.join(project_name, "resources", "logo", "logo.ico")
-        distpath    = os.path.join("deploy", "dist")
-        buildpath   = os.path.join("deploy", "build")
-        specpath    = os.path.join("deploy", "spec")
-
+def BuildOnLinux(programm_name, version_number):
+    cwd = os.getcwd()
+    specpath    = os.path.join(cwd, "deploy")
+    distpath    = os.path.join(specpath, "dist")
+    print('specpath: '+ specpath)
+    print('distpath: '+ distpath)
+    if  os.path.exists(specpath): 
+        os.chdir(specpath)
         Popen(["pyinstaller",
-               main_py,
-               "--onefile",
-               "--clean",
-               "--windowed",
-               "--distpath="+ distpath,
-               "--workpath="+ buildpath, 
-               "--specpath="+specpath,
-               "--icon=" + icon_file]).wait()
+                'main.spec']).wait()
 
-        file_name = project_name + "-windows-" + version_number + ".exe"
-        file_origin = os.path.join(distpath, "main.exe")
-        file_destination = os.path.join(distpath, file_name)
-
-        os.rename(file_origin, file_destination)
-
-    #note Code comes original from Hermann Krumrey: gitlab-build-scripts-0.6.0, but was deleted from his server
-    def BuildOnLinux(build_path, project_name, version_number):
-        Popen(["pyinstaller", os.path.join(project_name, "main.py"), "--onefile"]).wait()
-
-        file_name = project_name + "-linux-" + version_number
+        file_name = programm_name + "-linux-" + version_number
         file_origin = os.path.join("dist", "main")
-        file_destination = os.path.join(build_path, file_name)
-
+        file_destination = os.path.join(distpath, file_name)
+        if os.path.exists(file_destination): 
+            os.remove(file_destination)
         os.rename(file_origin, file_destination)
+        os.chdir(cwd)
+    else:
+        print('Path for deployment with spec file doesent exists!')
+        return False
 
-    def CreateSpec():
-       print('Create the installer spec')
 
-## Documentation for a method to build the executable with pyinstaller
-# Write the following code to some python script in the top directory and call it then from windows command line:
-# @code
-#from pylibcklb.installer.pyinstaller import build
-#
-#if __name__ == "__main__":
-#    name    = 'Example Program'
-#    version = '1.0.0' 
-#    build(name, version)
-# @endcode
-#   @param programm_name The name of the program
-#   @param version_number The version of the program
-def build(programm_name, version_number):
+def BuildOnWindows(programm_name, version_number):
     cwd = os.getcwd()
     specpath    = os.path.join(cwd, "deploy")
     distpath    = os.path.join(specpath, "dist")
@@ -107,6 +82,24 @@ def build(programm_name, version_number):
     else:
         print('Path for deployment with spec file doesent exists!')
         return False
+
+## Documentation for a method to build the executable with pyinstaller
+# Write the following code to some python script in the top directory and call it then from windows command line:
+# @code
+#from pylibcklb.installer.pyinstaller import build
+#
+#if __name__ == "__main__":
+#    name    = 'Example Program'
+#    version = '1.0.0' 
+#    build(name, version)
+# @endcode
+#   @param programm_name The name of the program
+#   @param version_number The version of the program
+def build(programm_name, version_number):
+    if platform.system() == 'Windows'
+        BuildOnWindows(programm_name, version_number)
+    else:
+        BuildOnLinux(programm_name, version_number)
